PR feedback: part 1
