+++ packages/ui-causal-graph-editor/src/graph-viewer/storybook/force-atlas.stories.tsx
@@ -16,10 +16,10 @@
  */
 import { Meta } from '@storybook/react';
 
-import { FRAUD } from '../../tests/mocks/graphs';
-import { ForceAtlasLayout } from '../shared/graph-layout';
-import { default as CausalGraphViewerComponent } from './causal-graph-editor';
-import { Template } from './utils/stories-utils';
+import { FRAUD } from '../../../tests/mocks/graphs';
+import { ForceAtlasLayout } from '../../shared/graph-layout';
+import { default as CausalGraphViewerComponent } from '../causal-graph-editor';
+import { Template } from './stories-utils';
 
 export default {
     component: CausalGraphViewerComponent,
