+++ packages/ui-causal-graph-editor/src/graph-viewer/storybook/marketing.stories.tsx
@@ -16,9 +16,9 @@
  */
 import { Meta } from '@storybook/react';
 
-import { MarketingLayout } from '../shared/graph-layout';
-import { default as CausalGraphViewerComponent } from './causal-graph-editor';
-import { Template, causalGraph, nodeTiersCausalGraph, nodeTiersList } from './utils/stories-utils';
+import { MarketingLayout } from '../../shared/graph-layout';
+import { default as CausalGraphViewerComponent } from '../causal-graph-editor';
+import { Template, causalGraph, nodeTiersCausalGraph, nodeTiersList } from './stories-utils';
 
 export default {
     component: CausalGraphViewerComponent,
