+++ packages/ui-causal-graph-editor/src/graph-viewer/storybook/planar.stories.tsx
@@ -17,11 +17,11 @@
 import { Meta } from '@storybook/react';
 import * as React from 'react';
 
-import { SHIPPED_UNITS } from '../../tests/mocks/graphs';
-import { PlanarLayout } from '../shared/graph-layout';
-import { LayeringAlgorithm } from '../shared/graph-layout/planar-layout';
-import { default as CausalGraphViewerComponent } from './causal-graph-editor';
-import { Template, causalGraph, nodeTiersCausalGraph, planarLayoutCausalGraph } from './utils/stories-utils';
+import { SHIPPED_UNITS } from '../../../tests/mocks/graphs';
+import { PlanarLayout } from '../../shared/graph-layout';
+import { LayeringAlgorithm } from '../../shared/graph-layout/planar-layout';
+import { default as CausalGraphViewerComponent } from '../causal-graph-editor';
+import { Template, causalGraph, nodeTiersCausalGraph, planarLayoutCausalGraph } from './stories-utils';
 
 export default {
     component: CausalGraphViewerComponent,
