+++ packages/ui-causal-graph-editor/src/shared/rendering/edge/definitions.tsx
@@ -40,7 +40,7 @@ export interface PixiEdgeStyle {
     /** Strength definition to be used if provided */
     strength?: EdgeStrengthDefinition;
     /** The number of collapsed edges at either end of the edge */
-    collapsedEdges?: number;
+    collapsedEdgesCount?: number;
     /** Current theme object */
     theme: DefaultTheme;
     /** Optional thickness override */
