+++ packages/ui-causal-graph-editor/src/shared/rendering/edge/edge-object.tsx
@@ -331,8 +331,8 @@ export class EdgeObject extends PIXI.utils.EventEmitter<(typeof MOUSE_EVENTS)[nu
 
         // Number symbols
         const edgeNumberSymbol = edgeSymbolsGfx.getChildByName<PIXI.Sprite>(EDGE_NUMBER_SYMBOL);
-        const numberSymbolTexture = textureCache.get(createKey(EDGE_NUMBER_SYMBOL, edgeStyle.collapsedEdges), () => {
-            if (edgeStyle.collapsedEdges === undefined) {
+        const numberSymbolTexture = textureCache.get(createKey(EDGE_NUMBER_SYMBOL, edgeStyle.collapsedEdgesCount), () => {
+            if (edgeStyle.collapsedEdgesCount === undefined) {
                 return new PIXI.Graphics();
             }
 
@@ -341,7 +341,7 @@ export class EdgeObject extends PIXI.utils.EventEmitter<(typeof MOUSE_EVENTS)[nu
                 fontSize: 18,
                 fill: colorToPixi(edgeStyle.color),
             });
-            const text = new PIXI.Text(edgeStyle.collapsedEdges, textStyle);
+            const text = new PIXI.Text(edgeStyle.collapsedEdgesCount, textStyle);
             return text;
         });
 
@@ -354,7 +354,7 @@ export class EdgeObject extends PIXI.utils.EventEmitter<(typeof MOUSE_EVENTS)[nu
                 (edgeGfx.rotation >= (-3 * Math.PI) / 2 && edgeGfx.rotation < -Math.PI)
             ) ?
                 -Math.PI / 2
-            :   Math.PI / 2;
+                : Math.PI / 2;
         [edgeStrengthSymbol.tint] = colorToPixi(edgeStyle.color);
         edgeNumberSymbol.alpha = 1;
